requirements.txt
Django==5.1.2
Pillow
django-unfold
django-import-export